requirements.txt
sqlalchemy==1.3.20
alembic==1.4.3
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
Flask-Migrate==2.5.3
psycopg2-binary==2.8.6
flask-cors==3.0.9
marshmallow==3.9.1
google-auth==1.23.0
requests==2.25.0